+++ src/mol-plugin-ui/controls/parameters.tsx
@@ -258,13 +258,7 @@ function renderSimple(options: { props: ParamProps<any>, state: { showHelp: bool
             title={desc}
             label={<>
                 {label}
-                {hasHelp &&
-                    <button className='msp-help msp-btn-link msp-btn-icon msp-control-group-expander' onClick={toggleHelp}
-                        title={desc || `${state.showHelp ? 'Hide' : 'Show'} help`}
-                        style={{ background: 'transparent', textAlign: 'left', padding: '0' }}>
-                        <Icon svg={HelpOutlineSvg} />
-                    </button>
-                }
+                {hasHelp && <ToggleParamHelpButton show={state.showHelp} toggle={toggleHelp} title={desc} />}
             </>}
             control={control}
         />
@@ -273,6 +267,14 @@ function renderSimple(options: { props: ParamProps<any>, state: { showHelp: bool
         </div>}
         {addOn}
     </>;
+}
+
+export function ToggleParamHelpButton({ show, toggle, title }: { show: boolean, toggle: () => void, title?: string }) {
+    return <button className='msp-help msp-btn-link msp-btn-icon msp-control-group-expander' onClick={toggle}
+        title={title || `${show ? 'Hide' : 'Show'} help`}
+        style={{ background: 'transparent', textAlign: 'left', padding: '0' }}>
+        <Icon svg={HelpOutlineSvg} />
+    </button>;
 }
 
 export abstract class SimpleParam<P extends PD.Any> extends React.PureComponent<ParamProps<P>, { showHelp: boolean }> {